MssubMcp (mssub_mcp v0.1.0)

MssubMcp - Global Systems Management

API for the Master Control Program Subsystem.

The Master Control Program Subsystem provides global services such as tenant and Application Instance management, global user authentication, global system configuration services, and a global enumeration service.

Functions

process_operation(operation)

@spec process_operation((() -> any())) :: any()

Processes the given function in the context of the MCP services & Datastore.

Returns the return value of the provided function.

Parameters

operation - a function which wraps the operations to be executed in the MCP service context.

Examples

Retrieving an Msdata.SystOwners record from the MCP database.

iex> mcp_operation = fn -> MscmpSystInstance.get_owner_by_name("owner1") end
iex> {:ok, %Msdata.SystOwners{}} = MssubMcp.process_operation(mcp_operation)